Converted all numbers to int. Refactored api routes to make more logical sense for mediaItems. Changed all times to store seconds instead of minutes. Added new routes to get media item resume times.

+++ PlumMediaCenter/Business/Data/Movie.cs
@@ -2,7 +2,7 @@ namespace PlumMediaCenter.Data
 {
     public class Movie
     {
-        public ulong? Id;
+        public int? Id;
         public string Title;
         public string Summary;
         public string Description;
