+++ PlumMediaCenter/Business/Data/Source.cs
@@ -4,7 +4,7 @@ namespace PlumMediaCenter.Data
 {
     public class Source
     {
-        public ulong? Id;
+        public int? Id;
         public string FolderPath;
         /// <summary>
         /// The type of media this source contains (movies, series)
